nanobody_design/requirements-nanobody.txt
biopython==1.82
torch==2.4.1
transformers==4.44.2
